finish refactor db
