+++ servicelog.api/database/db.py
@@ -1,6 +1,6 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from base import Base
+from database.base import Base
 
 DATABASE_URL = "sqlite:///./servicelog.db"
 
@@ -8,8 +8,12 @@ engine = create_engine(DATABASE_URL)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
 def init_db():
-    import models.user
-    import models.customer
+    from models.user import User
+    from models.customer import Customer
+    from models.machine_model import MachineModel
+
+    print("Registered tables:", Base.metadata.tables.keys())
+
     Base.metadata.create_all(bind=engine)
 
 init_db()
