+++ servicelog.api/models/__init__.py
@@ -1,4 +1,3 @@
 from .user import User
 from .customer import Customer
 from .machine_model import MachineModel
-from enums import MachineTypes
